backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
boto3==1.34.0
cryptography==42.0.5
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
